test: remove _new suffix from test files
