+++ tests/model_connectors/test_district_heating_system.py
@@ -42,17 +42,20 @@ from geojson_modelica_translator.model_connectors.couplings.graph import (
 from geojson_modelica_translator.model_connectors.districts.district import (
     District
 )
-from geojson_modelica_translator.model_connectors.energy_transfer_systems.cooling_indirect import (
-    CoolingIndirect
+from geojson_modelica_translator.model_connectors.energy_transfer_systems.ets_cold_water_stub import (
+    EtsColdWaterStub
 )
-from geojson_modelica_translator.model_connectors.energy_transfer_systems.ets_hot_water_stub import (
-    EtsHotWaterStub
+from geojson_modelica_translator.model_connectors.energy_transfer_systems.heating_indirect import (
+    HeatingIndirect
 )
 from geojson_modelica_translator.model_connectors.load_connectors.time_series import (
     TimeSeries
 )
-from geojson_modelica_translator.model_connectors.networks.network_chilled_water_stub import (
-    NetworkChilledWaterStub
+from geojson_modelica_translator.model_connectors.networks.network_2_pipe import (
+    Network2Pipe
+)
+from geojson_modelica_translator.model_connectors.plants.heating_plant import (
+    HeatingPlant
 )
 from geojson_modelica_translator.system_parameters.system_parameters import (
     SystemParameters
@@ -61,9 +64,9 @@ from geojson_modelica_translator.system_parameters.system_parameters import (
 from ..base_test_case import TestCaseBase
 
 
-class DistrictSystemTest(TestCaseBase):
-    def test_district_system(self):
-        project_name = "district_system_new"
+class DistrictHeatingSystemNewTest(TestCaseBase):
+    def test_district_heating_system(self):
+        project_name = 'district_heating_system'
         self.data_dir, self.output_dir = self.set_up(os.path.dirname(__file__), project_name)
 
         # load in the example geojson with a single office building
@@ -74,24 +77,24 @@ class DistrictSystemTest(TestCaseBase):
         filename = os.path.join(self.data_dir, "time_series_system_params_ets.json")
         sys_params = SystemParameters(filename)
 
-        # Create the time series load, ets and their coupling
-        time_series_load = TimeSeries(sys_params, self.gj.json_loads[0])
-        cooling_indirect_system = CoolingIndirect(sys_params)
-        ts_ci_coupling = Coupling(time_series_load, cooling_indirect_system)
+        # create network and plant
+        network = Network2Pipe(sys_params)
+        heating_plant = HeatingPlant(sys_params)
 
-        # create chilled water stub for the ets
-        chilled_water_stub = NetworkChilledWaterStub(sys_params)
-        ci_cw_coupling = Coupling(cooling_indirect_system, chilled_water_stub)
+        # create our our load/ets/stubs
+        all_couplings = [
+            Coupling(network, heating_plant)
+        ]
+        for _ in range(6):
+            time_series_load = TimeSeries(sys_params, self.gj.json_loads[0])
+            heating_indirect_system = HeatingIndirect(sys_params)
+            cold_water_stub = EtsColdWaterStub(sys_params)
+            all_couplings.append(Coupling(time_series_load, heating_indirect_system))
+            all_couplings.append(Coupling(time_series_load, cold_water_stub))
+            all_couplings.append(Coupling(heating_indirect_system, network))
 
-        #  create hot water stub for the load
-        hot_water_stub = EtsHotWaterStub(sys_params)
-        ts_hw_coupling = Coupling(time_series_load, hot_water_stub)
-
-        graph = CouplingGraph([
-            ts_ci_coupling,
-            ci_cw_coupling,
-            ts_hw_coupling,
-        ])
+        # create the couplings and graph
+        graph = CouplingGraph(all_couplings)
 
         district = District(
             root_dir=self.output_dir,
